# [Phase 3] Performance Optimization