--- a/frontend/vite.config.ts
+++ b/frontend/vite.config.ts
@@ -7,8 +7,27 @@ export default defineConfig({
   build: {
     outDir: 'dist',
     emptyOutDir: true,
+    rollupOptions: {
+      output: {
+        manualChunks: {
+          vendor: ['react', 'react-dom'],
+          router: ['react-router-dom'],
+          ui: ['lucide-react']
+        }
+      }
+    },
+    minify: 'terser',
+    target: 'es2020',
+    chunkSizeWarningLimit: 1000
   },
   server: {
     port: 3000,
   },
+  optimizeDeps: {
+    include: ['react', 'react-dom'],
+    exclude: ['@wailsapp/runtime']
+  },
+  define: {
+    __DEV__: JSON.stringify(false)
+  }
 })