--- a/internal/database/db.go
+++ b/internal/database/db.go
@@ -25,6 +25,11 @@ func Open(dbPath string) (*DB, error) {
 
 	db := &DB{conn: conn}
 	
+	// Optimize database connection for performance
+	if err := db.optimize(); err != nil {
+		return nil, fmt.Errorf("failed to optimize database: %w", err)
+	}
+	
 	// Run migrations
 	if err := db.migrate(); err != nil {
 		return nil, fmt.Errorf("failed to migrate database: %w", err)
@@ -42,6 +47,26 @@ func (db *DB) Close() error {
 }
 
 // migrate is implemented in migrations.go
+
+// optimize configures SQLite for better performance
+func (db *DB) optimize() error {
+	optimizations := []string{
+		"PRAGMA journal_mode=WAL",        // Write-Ahead Logging for better concurrency
+		"PRAGMA synchronous=NORMAL",      // Faster writes with reasonable safety
+		"PRAGMA cache_size=10000",        // 10MB cache (default is 2MB)
+		"PRAGMA temp_store=MEMORY",       // Keep temp tables in memory
+		"PRAGMA mmap_size=268435456",     // 256MB memory map size
+		"PRAGMA optimize",               // Optimize query planner
+	}
+	
+	for _, pragma := range optimizations {
+		if _, err := db.conn.Exec(pragma); err != nil {
+			return fmt.Errorf("failed to execute optimization '%s': %w", pragma, err)
+		}
+	}
+	
+	return nil
+}
 
 // GetConn returns the underlying database connection
 func (db *DB) GetConn() *sql.DB {