… CLAUDE.md
… PERFORMANCE_REPORT.md
… internal/app/performance.go
… internal/parser/service.go